fastapi-app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
pandas>=2.0.0
openpyxl>=3.1.0   # нужен для чтения .xlsx
xlrd>=2.0.1       # если вдруг будут старые .xls
bcrypt==3.2.2
passlib[bcrypt]==1.7.4
email-validator

